Unary Operations for MDMatrix works - can now perform operations on a Matrix of Objects that implement SupportsBinaryOperations. Requries work on displaying and only rounding (in complexNubmers) where needed. Also, test MDMatrix operations of different dimensions

+++ Matrices/src/MDEntities/MatrixDouble.java
@@ -77,7 +77,7 @@ public class MatrixDouble implements Equatable, SupportsBinaryOperations<MatrixD
         }
 
         private static double round(double value){
-            return Math.round(value * 100000) / 100000;
+            return (double)Math.round(value * 100000) / 100000.0;
         }
     }
